requirements.txt
fastapi
uvicorn
requests
python-dotenv
jinja2
pydantic
python-multipart
vonage>=4.1.0
vonage-messages
